+++ java/src/hype/core/drawable/HDrawable.java
@@ -36,13 +36,13 @@ import processing.core.PVector;
 public abstract class HDrawable extends HNode<HDrawable>
 		implements HDirectable, HHittable {
 	
-	/** TODO */
+	public static final int DEFAULT_FILL = 0xFFFFFFFF;
+	public static final int DEFAULT_STROKE = 0xFF000000;
+	public static final int DEFAULT_WIDTH = 100;
+	public static final int DEFAULT_HEIGHT = 100;
 	public static final byte BITMASK_PROPORTIONAL = 1;		// 0b0001
-	/** TODO */
 	public static final byte BITMASK_TRANSFORMS_CHILDREN = 2;// 0b0010
-	/** TODO */
 	public static final byte BITMASK_STYLES_CHILDREN = 4;	// 0b0100
-	/** TODO */
 	public static final byte BITMASK_ROTATES_CHILDREN = 8;	// 0b1000
 	
 	/** The parent of this drawable @see parent() */
@@ -110,14 +110,14 @@ public abstract class HDrawable extends HNode<HDrawable>
 	public HDrawable() {
 		_alphaPerc = 1;
 		
-		_fill = HConstants.DEFAULT_FILL;
-		_stroke = HConstants.DEFAULT_STROKE;
+		_fill = DEFAULT_FILL;
+		_stroke = DEFAULT_STROKE;
 		_strokeCap = PConstants.ROUND;
 		_strokeJoin = PConstants.MITER;
 		_strokeWeight = 1;
 		
-		_width = HConstants.DEFAULT_WIDTH;
-		_height = HConstants.DEFAULT_HEIGHT;
+		_width = DEFAULT_WIDTH;
+		_height = DEFAULT_HEIGHT;
 	}
 	
 	/**